// sanity.types.ts
export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;
// Source: ./sanity/lib/post/getPost.ts
// Variable: getPostQuery
// Query: *[_type == "posts" && _id == $id][0]{  ...,  "comments": *[_type == "comment" && post._ref == ^._id] | order(_createdAt desc)}
export type GetPostQueryResult = {
  _id: string;
  _type: "posts";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  title?: string;
  body?: Array<{
    children?: Array<{
      marks?: Array<string>;
      text?: string;
      _type: "span";
      _key: string;
    }>;
    style?: "blockquote" | "h1" | "h2" | "h3" | "h4" | "h5" | "h6" | "normal";
    listItem?: "bullet" | "number";
    markDefs?: Array<{
      href?: string;
      _type: "link";
      _key: string;
    }>;
    level?: number;
    _type: "block";
    _key: string;
  }>;
  tierAccess?: "backstage" | "crew" | "vip";
  coverImage?: {
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
    };
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    alt?: string;
    _type: "image";
  };
  comments: Array<{
    _id: string;
    _type: "comment";
    _createdAt: string;
    _updatedAt: string;
    _rev: string;
    post?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "posts";
    };
    name?: string;
    userImageUrl?: string;
    email?: string;
    comment?: string;
  }>;
} | null;